test(deepkit): try to skip in nodejs 12.x
